requirements.txt
Django==1.10.6
factory-boy==2.8.1
Faker==0.7.10
freezegun==0.3.8
Markdown==2.6.8
psycopg2==2.7.1
python-dateutil==2.6.0
six==1.10.0
wheel==0.24.0
